Remove PlayerController script, add Stag prefab
The PlayerController.cs script was removed since the player character's movement is now handled by a new system. This change will improve the overall gameplay experience.

The new PF_Stag prefab was introduced to enhance the forest environment in the game. This new asset will be used to make the