Add game scoring and projectile firing features
Implemented a scoring system where the score increases when the projectile collides with an animal (reference added in the Projectile script). Added a script to instantiate a projectile from the player when the space bar is pressed (inside the PlayerController script). An additional delay between firing projectiles was also implemented. Tags for various animal prefabs were set to "Animal" for detection. SpawnManager script was also modified to allow animals to spawn from different directions.