"Update properties and components of prefabs and scripts"
A series of modifications were made affecting different prefabs and scripts.
In the horse and cow prefabs, component files were replaced and BoxCollider properties were adjusted, likely to result in better interactions with objects or environment. The top and lower boundaries in the DestroyOutOfBounds.cs script were changed from variables to constants, with added conditions for object deletion which may help in memory and performance optimization. Also the BonusFeatures.unity scene, prefab instances and associated properties were modified. It seems a prefab was moved within the scene hierarchy and its local scale adjusted in the PF_BFStag.prefab.